+++ erp/apps/api/src/Erp.UseCases/Employees/SetLeaveQuota/SetLeaveQuotaCommand.cs
@@ -3,4 +3,4 @@ using Erp.UseCases.Common;
 namespace Erp.UseCases.Employees.SetLeaveQuota;
 
 /// <summary>Null <paramref name="Days"/> clears the override for that type.</summary>
-public sealed record SetLeaveQuotaCommand(Guid EmployeeId, string Type, int? Days, Caller Caller);
+public sealed record SetLeaveQuotaCommand(Guid EmployeeId, string Type, decimal? Days, Caller Caller);
